add authorization to edit and delete functionality

+++ rest-apis/controllers/feed.js
@@ -128,6 +128,12 @@ exports.updatePost = (req, res, next) => {
                 throw error;  
             } 
 
+            if(post.creator.toString() !== req.userId){
+                const error = new Error('Not Authorized');
+                error.statusCode = 403;
+                throw error;
+            }
+
             if(imageUrl !== post.imageUrl){
                 clearImage(post.imageUrl);
             }
@@ -160,7 +166,13 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;  
             } 
-            // check logged in user
+
+            if(post.creator.toString() !== req.userId){
+                const error = new Error('Not Authorized');
+                error.statusCode = 403;
+                throw error;
+            }
+            
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId)
         })
